Added Database Table import and export; had to update DB to version 2.
Save all data before starting the server (by changing the DB version to 1 in Java, using the new export function and changing it back afterward, then using the import function).
Three files are provided for the import function. (Can be called via URL or in the ImportExportRequest-class, as long as the server is running).